trader: extend server to subscribe to block notifications
Some of the trader's subsystems will require the need for knowing what
the best height of the chain is. Rather than having each subsystem
subscribe for their own block notifications, we decide to do so only
once within the main server and have it provide the best height to its
subsystems.