chaninfo: add new package to retrieve channel info for auctioneer
When traders create channels as part of a batch, the auctioneer provides
a guarantee to punish any traders who violate the lifetime of a channel
(also known as their maturity height). In order for the auctioneer to
carry out this duty, it'll need to retrieve a set of channel keys from
each trader to detect, in the event of a lifetime violation, who the
responsible trader was. This channel info is retrieved by the traders
from their encrypted SCBs.