clmrpc: add RelevantBatchSnapshot RPC
This RPC exposes batch snapshots to traders. This is useful when the
trader wants to receive information about a batch they've participated
in for the relevant accounts.

Alternatively, this RPC can also be used by traders as a recovery
mechanism to determine whether they've made it into a finalized batch by
comparing their known batch transaction with the auctioneer's. This will
be expanded on in a later commit.