client/rpc: fix panic in sendAcceptBatch
In this commit, we fix a panic in `sendAcceptBatch` due to the fact that
the method as is, creates a slice with enough _capacity_ but a zero
length. In this case, trying to access an index will result in a panic.
In this commit, we shift things over to use append instead of index into
the slice.