clmrpc: add expiry and trader key to account reservation
To be able to recover an account that was only reserved before the
data loss occurred at the trader, we need to store more information
in the auctioneer's reservation. Specifically we also need to store
the account expiry, the trader key and the value. Previously the
value was only validated, now it will also be stored in the
reservation.