clmrpc: add error type for reserved account
If we only have a reservation on the auctioneer's side when a
trader tries to connect, we cannot create a full account subscription
and therefore not follow the normal message flow for recovery. Instead
we return a special error message that contains the information we have
that should be enough to restore the account on the trader's side.

The trader still has to provide a valid signature with the trader key
so no probing for this information should be possible by any third
party.