fundingmgr: let bidder connect to asker
To allow a bidder to be a mobile or Tor only node, we only require the
asker to be reachable through clearnet. This requirement wasn't
implemented correctly in the funding flow. This commit fixes that and
turns the logic around. The asker creates the channel but the
bidder does all the work for it (connecting out and registering the
funding shim).