fundingmgr: don't block on peer connection attempt
Since the call to connect to a peer is no longer async, a failed
connection would block and stall the whole execution logic until the
setupCtx was canceled. At that point the node we failed to connect with
would not be communicated back to the auctioneer, and we could get
matched with the same node in the next batch attempt.

This fixes that by starting the connection attempt in a go routine.