mod+macaroons: bump to lnd v0.11.1-beta.rc3
To make sure we can use the latest version of pool in LiT, we need to
make sure pool compiles against the actual v0.11.1 branch of lnd. There
were some conflicting changes to the macaroon service that didn't make
it to the branch that we need to roll back. Those changes don't affect
pool at all as we were always using the default root key ID anyway.