fundingmgr_test: use manual peer list only, wait for connections
Because connectToMatchedTrader is now called in a goroutine, we cannot
rely on the connections already being established when we enter
waitForPeerConnection. So we switch the test to not take the list of
connected peers from the connection attempts but instead use a manual
map of peers we expect to be connected.
We then wait a bit for all connection establishment calls to finish to
make sure the calls are performed in the background.