poolrpc: move towards set duration buckets instead of min/max duration
In this commit, we make a series of RPC changes to do away with the
min/max order preference and instead replace that with a single unified
field for the duration of a lease. The Terms method has been extended to
return the current set of active leases. Things have been done in this
way to allow us to announce a new duration, accept orders, but not yet
enable market clearing for it.

The changes in this commit were intended to be non-breaking, which is
why the Bid/Ask message distinction still exists, even though with this
commit they nearly share all the same fields. In the future further
fields may be added that still make both messages asymmetric.