auctioneer: add own quit channel to subscription
If the RPC server receives an error on the stream error channel, it will
issue a reconnect to the auctioneer. If any account was still in the
process of authenticating, this will cause it to bubble up an error and
initiate a shutdown of the trader instead. If something goes wrong in
the authentication part, it's usually non-recoverable (unless the server
is shutting down at that time, in which case the user has to try again
manually at a later point in time).