auctioneer: try connecting with sync RPC
To make the auctioneer connection retry mechanism a bit more robust, we
try a "cheap" (resource wise) synchronous RPC call instead of the
streaming SubscribeBatchAuction call that'll only fail once we actually
try to authenticate an account.
That way we can catch errors if the auctioneer server is down and try a
proper reconnect.