mod: make auctioneerrpc its own sub-module
To make it possible to replace the package in the server integration
tests, we declare the auctioneerrpc package as its own submodule. That
way we can just plug in the server version in the itest if we have changes in the
server proto that aren't merged to the client repo yet.