make: fix ldflags to include commit in version string
The Makefile contained a number of copy/paste errors that we fix with
this commit: The make_ldflags function and RELEASE_TAGS variable were
missing and therefore resulted in empty strings.
With those things removed, we can then go ahead and build the correct
LDFLAGS variable and actually use it when building or installing the
binaries.