poolrpc+rpcserver: add GetInfo and StopDaemon RPCs
To give developers and users more information about the current state of
the trader daemon, we add a new GetInfo RPC that shows a summary of all
accounts and orders known to the system.
And because there currently is no other way than to <ctrl>+c or kill the
daemon, we add a StopDaemon RPC as well for proper shutdown.