sidecar: allow recipient ID to be derived using base offer ticket
In this commit, we fix a bug in the retransmission case for the
provider. Before this commit, the provider would attempt to re-send the
initial offer ticket on restart if it hadn't yet received the registered
ticket. This fails as the recipient's stream ID is derived from their
pubkey information that's only contained in the registered ticket.

To fix this, we'll simply use the sha512 hash of the offer sig since
it's known when the ticket is initially created.