gomod: bump to lndclient 13-7
This commit bumps pool to lndclient 13-7 so that code that depends
on lnd 0.13 can import pool. As is, pool is running with lndclient 0.12,
but pins minimum version to 0.11, so we are already in a state where
some of the apis available at compile time are not enforced by minimum
run version.

This commit bumps this minimum version up to 0.12 to minimize the
difference between the compile and runtime lnd version.