auctioneer: decrease the `reconnectRetries` value
Whenever the client is not able to connect to the server it waits some
backoff time and retries it. In some cases that fixes the the problem, for
example if the backend was down for some reason. However, if the client is
sending an invalid request again and again those retries will pollute
our server logs.

Set a more reasonable `reconnectRetries` value (90) so clients don't hang
sending invalid requests "forever" (65535)