account/manager: fix np2wkh deposits
The `.UnsignedTx` of a psbt does not include the related
SignatureScripts. Because SignatureScripts are not part of the signed
data, the partial signature for a psbt that does not include them is a
valid signature. However, the `TxHash` will change between a tx that has
SignatureScripts set and one that does not.

This behaviour led the auctioneer and the client to record the wrong
outpoint when using np2wkh.