script: add `verify-install.sh` for signed releases
The process is similar to the one used in LND. The repo includes a list
of keys that can be used to sign a valid releases.

By now, we need a signature of 2 out of 4 to mark the release as valid. This
number can increase when more keys are added to the list.