cmd/pool: increase default number of keys to try in recovery
This commit fixes an issue where a large number of attempts at creating
an account lead to the index of the first successful account to be
higher than the previous number of keys we tried by default. This
resulted in the account not being recovered without manually overriding
this value in the command line.