make+tools: don't compile with build tag to avoid collision
It looks like one of the projects pulled in by the linter also uses the
"tools" build tag. When we turn on that build tag, we get build errors
in github.com/ryancurrah/gomodguard. So we just remove the build tag
during compilation as it's not actually needed (the build tag is just
there to not pull in the dependencies during the normal build).